Refactoring, splitting code in multiple files and implementing the MultiDeposit case
