+++ src/main/scala/nl/knaw/dans/easy/ingest_flow/MultiDepositExecution.scala
@@ -1,0 +1,25 @@
+package nl.knaw.dans.easy.ingest_flow
+
+import scala.util.Try
+
+object MultiDepositExecution extends Execution {
+
+  def run(implicit settings: Settings): Try[String] = {
+    for {
+      urn <- requestUrn()
+      doi <- requestDoi()
+      _ <- stageDataset(urn, doi, false) // TODO refactor?
+      pidDictionary <- ingestDataset()
+      datasetPid <- getDatasetPid(pidDictionary)
+      _ <- waitForFedoraSync(datasetPid, pidDictionary)
+      _ <- updateFsRdb(datasetPid)
+      _ <- updateSolr(datasetPid)
+      _ <- deleteSdoSetDir()
+      _ <- setDepositStateToArchived(datasetPid)
+      _ <- deleteBag()
+      _ <- deleteGitRepo()
+    } yield datasetPid
+  }
+
+  def requestDoi()(implicit s: Settings): Try[String] = requestPid("doi")
+}
