+++ src/main/scala/nl/knaw/dans/easy/ingest_flow/package.scala
@@ -15,12 +15,13 @@
   */
 package nl.knaw.dans.easy
 
-import java.io.File
+import java.io.{File, IOException}
 import java.net.URL
 import java.util.Properties
 
 import com.yourmediashelf.fedora.client.FedoraCredentials
 import org.apache.commons.configuration.PropertiesConfiguration
+import org.apache.commons.io.FileUtils
 
 import scala.util.{Failure, Success, Try}
 
@@ -29,21 +30,21 @@ package object ingest_flow {
   // types
 
   // case classes
-  case class Settings(storageUser: String,
-                      storagePassword: String,
-                      storageServiceUrl: URL,
-                      fedoraCredentials: FedoraCredentials,
-                      numSyncTries: Int,
-                      syncDelay: Long,
-                      ownerId: String,
-                      datasetAccessBaseUrl: String,
-                      depositDir: File,
-                      checkInterval: Int,
-                      maxCheckCount: Int,
-                      sdoSetDir: File,
-                      postgresURL: String,
-                      solr: String,
-                      pidgen: String)
+  case class Settings(storageUser: String = null,
+                      storagePassword: String = null,
+                      storageServiceUrl: URL = null,
+                      fedoraCredentials: FedoraCredentials = null,
+                      numSyncTries: Int = 0,
+                      syncDelay: Long = 0,
+                      ownerId: String = null,
+                      datasetAccessBaseUrl: String = null,
+                      depositDir: File = null,
+                      checkInterval: Int = 0,
+                      maxCheckCount: Int = 0,
+                      sdoSetDir: File = null,
+                      postgresURL: String = null,
+                      solr: String = null,
+                      pidgen: String = null)
 
   object Version {
     def apply(): String = {
@@ -67,11 +68,27 @@ package object ingest_flow {
     }
   }
 
+  implicit class FileExtensions(val file: File) extends AnyVal {
+    /**
+      * Deletes a directory recursively.
+      *
+      * @throws IOException in case deletion is unsuccessful
+      */
+    def deleteDirectory() = FileUtils.deleteDirectory(file)
+  }
+
   def getUserId(depositDir: File): String = {
     new PropertiesConfiguration(new File(depositDir, "deposit.properties")).getString("depositor.userId")
   }
 
   def getBagDir(depositDir: File): Option[File] = {
     depositDir.listFiles.find(f => f.isDirectory && f.getName != ".git")
+  }
+
+  def setDepositState(state: String, description: String)(implicit s: Settings): Try[Unit] = Try {
+    val stateFile = new PropertiesConfiguration(new File(s.depositDir, "deposit.properties"))
+    stateFile.setProperty("state.label", state)
+    stateFile.setProperty("state.description", description)
+    stateFile.save()
   }
 }
